Flyttet fagdag til en page istedenfor md
