+++ src/components/fagdag.less
@@ -1,5 +1,7 @@
 
 .sb1-fagdag-page{
+    font-size: 24px;
+    font-family: 'MuseoSans-300', arial, sans-serif;
     &__header{
         background-color: @ffe-blue-ice;
     } 
@@ -38,7 +40,7 @@
         animation-name: colorChange1;
         animation-duration: 2s; 
         animation-timing-function: ease-out; 
-        animation-delay: 0s;
+        animation-delay: 4s;
         animation-iteration-count: 1;
     }
     @keyframes colorChange1 {
@@ -52,7 +54,7 @@
         animation-name: colorChange2;
         animation-duration: 2s; 
         animation-timing-function: ease-out; 
-        animation-delay: 1.5s;
+        animation-delay: 5.5s;
         animation-iteration-count: 1;
     }
     @keyframes colorChange2 {
@@ -66,7 +68,7 @@
         animation-name: colorChange3;
         animation-duration: 2s; 
         animation-timing-function: ease-out; 
-        animation-delay: 3s;
+        animation-delay: 7s;
         animation-iteration-count: 1;
     }
     @keyframes colorChange3 {
@@ -80,7 +82,7 @@
         animation-name: colorChange4;
         animation-duration: 2s; 
         animation-timing-function: ease-out; 
-        animation-delay: 5.5s;
+        animation-delay: 8.5s;
         animation-iteration-count: 1;
     }
     @keyframes colorChange4 {
@@ -92,11 +94,11 @@
 
 
     .typewriter{
-        max-width: 460px;
+        max-width: 730px;
 
         p {
             font-family: monospace;
-            font-size: 14px;
+            font-size: 22px;
             overflow: hidden;
             border-right: .15em solid @ffe-blue-royal;
             white-space: nowrap;
