--- src/pages/static/fagdag.md
@@ -1,24 +1,0 @@
----
-path: "/fagdag"
-title: "Fagdag"
-subtitle: "Hørt om fagdagen?"
-pagetype: "fagdag"
-image: "./img/fagdag.jpg"
-description: "For våre utviklere og designere er én dag i uken dedikert til faglig utvikling. Det som startet som et eksperiment i 2018, er nå en fast aktivitet. Her kan du lese litt mer om hva vi driver med på ukens beste dag!"
----
-
-## “Torsdag er den nye lørdagen!”
-
-Utviklere og designere motiveres av læring. Det gjelder selvfølgelig også hos oss. Undersøkelser vi har gjort blant utviklere viser at det aller viktigste for både studenter, ferske utviklere og "gamle" utviklere er nettopp læring. Du kjenner deg vel igjen?
-
-For å sikre kontinuerlig læring har vi satt av hver torsdag til fag. Vi kaller denne dagen *Fagdag*. Denne dagen forlater vi det trygge teamet vårt, og samler oss i nye selvorganiserende grupperinger. Vi bruker dagene på kursing, vi koder nye løsninger, vi lager foredrag og vi gjør forbedringer på fellesløsninger.
-
-## “Kultur spiser strategi til frokost”
-På hver konferanse dukker sistatet “Kultur spiser strategi til frokost” opp. Det er ikke uten grunn, og vi bruker Fagdagen til å bygge kultur. Gjennom at vi samles en dag uken for å utlukkende å fokusere på fag, så får vi mulighet til å utvikle en felles SpareBank 1 kultur. Vi ønsker at denne kulturen preges av åpenhet, trygghet og samhold.
-
-## Open Source
-Noen av fagdagene er dedikert til KodeKatas. Oppgavene finner vi på http://codingdojo.org/
-
-Løsningene legger vi selvfølgelig ut Open Source. Ta en titt på https://github.com/SpareBank1/fagdagen
-
-Kan vi friste med Game of Life implementert i Basic (for kjøring på Commodore 64): https://github.com/SpareBank1/fagdagen/tree/master/KodeKata/game-of-life/C64-Basic
